src/containers/image-pii-detection/requirements.txt
requests
boto3
six==1.16.0
opencv-python-headless==4.5.3.56
numpy<=1.23.5
onnxruntime
Pillow==8.4.0
pyclipper==1.3.0
Shapely==1.7.1
base64image==0.5.1
urllib3==1.26.6
python-dateutil==2.8.2
certifi==2022.12.7
idna==2.10
chardet==4.0.0
pypdf==3.12.1
python-magic==0.4.27
python-docx==0.8.11
bs4==0.0.1
pandas==1.5.3